--- Bisal/codes/log.txt
@@ -1,18 +1,0 @@
-[2024-07-05 10:11:16] This is a log message
-[2024-07-05 10:11:16] Another log entry
-[2024-07-05 10:41:27] This is a log message
-[2024-07-05 10:41:27] Another log entry
-[2024-07-05 10:45:23] This is a log message
-[2024-07-05 10:45:23] Another log entry
-[2024-07-05 10:58:11] This is a log message
-[2024-07-05 10:58:11] Another log entry
-[2024-07-05 10:58:59] This is a log message
-[2024-07-05 10:58:59] Another log entry
-[2024-07-05 11:21:11] This is a log message
-[2024-07-05 11:21:11] Another log entry
-[2024-07-05 11:23:30] This is a log message
-[2024-07-05 11:23:30] Another log entry
-[2024-07-05 11:30:10] This is a log message
-[2024-07-05 11:30:10] Another log entry
-[2024-07-05 11:31:11] This is a log message
-[2024-07-05 11:31:11] Another log entry
