--- Bisal/codes/student.txt
@@ -1,5 +1,0 @@
-
-Student Information:
-Name: aisha
-ID: 345
-Grades: 45.000000 67.000000 89.000000
